ALL METHODS CREATED
Add all methods and tested them with and without a block given
